requirements.txt
tqdm
numpy
opencv-python
scipy
visdom
torch
torchvision
pyyaml
matplotlib
pyamg
